requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
msgpack-python==0.5.6
packaging==23.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
redis==4.5.4
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
typing==3.7.4.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
